requirements.txt
Django==5.1.6
python-dotenv==1.0.1 
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0 
djangorestframework==3.16.0
python-dateutil==2.9.0.post0
six==1.17.0